refactor : OrderTable 주문 테이블 관련 패키지 분리
